+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## unreleased
 ### Added
 * Added rendering support for layered fur shell rendering to xc3_wgpu.
+* Added rendering support for selecting texture coordinate attributes to xc3_wgpu.
 
 ### Fixed
 * Fixed an issue where texture layers would only assign the first channel.